Improve folder structure.
