+++ cmake/Doxyfile.in
@@ -795,6 +795,7 @@ INPUT = "@CMAKE_SOURCE_DIR@/gqcp/include" \
         "@CMAKE_SOURCE_DIR@/gqcp/include/Basis" \
         "@CMAKE_SOURCE_DIR@/gqcp/include/Basis/Integrals" \
         "@CMAKE_SOURCE_DIR@/gqcp/include/Basis/Integrals/Interfaces" \
+        "@CMAKE_SOURCE_DIR@/gqcp/include/Basis/Integrals/Primitive" \
         "@CMAKE_SOURCE_DIR@/gqcp/include/Basis/MullikenPartitioning" \
         "@CMAKE_SOURCE_DIR@/gqcp/include/Basis/ScalarBasis" \
         "@CMAKE_SOURCE_DIR@/gqcp/include/Basis/SpinorBasis" \
@@ -841,6 +842,7 @@ INPUT = "@CMAKE_SOURCE_DIR@/gqcp/include" \
         "@CMAKE_SOURCE_DIR@/gqcp/src/Basis" \
         "@CMAKE_SOURCE_DIR@/gqcp/src/Basis/Integrals" \
         "@CMAKE_SOURCE_DIR@/gqcp/src/Basis/Integrals/Interfaces" \
+        "@CMAKE_SOURCE_DIR@/gqcp/src/Basis/Integrals/Primitive" \
         "@CMAKE_SOURCE_DIR@/gqcp/src/Basis/ScalarBasis" \
         "@CMAKE_SOURCE_DIR@/gqcp/src/Basis/SpinorBasis" \
         "@CMAKE_SOURCE_DIR@/gqcp/src/Basis/Transformations" \
