+++ gqcp/include/Basis/Integrals/IntegralEngine.hpp
@@ -23,13 +23,13 @@
 #include "Basis/Integrals/Interfaces/LibintOneElectronIntegralEngine.hpp"
 #include "Basis/Integrals/Interfaces/LibintTwoElectronIntegralEngine.hpp"
 #include "Basis/Integrals/OneElectronIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveAngularMomentumIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveCanonicalKineticEnergyIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveCoulombRepulsionIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveDipoleIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveLinearMomentumIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveNuclearAttractionIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveOverlapIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveAngularMomentumIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveCanonicalKineticEnergyIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveCoulombRepulsionIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveDipoleIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveLinearMomentumIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveNuclearAttractionIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveOverlapIntegralEngine.hpp"
 #include "Basis/Integrals/TwoElectronIntegralEngine.hpp"
 #include "Operator/FirstQuantized/Operator.hpp"
 #include "Utilities/aliases.hpp"
