+++ gqcp/include/gqcp.hpp
@@ -22,9 +22,6 @@
 #include "Basis/Integrals/BaseOneElectronIntegralEngine.hpp"
 #include "Basis/Integrals/BaseTwoElectronIntegralBuffer.hpp"
 #include "Basis/Integrals/BaseTwoElectronIntegralEngine.hpp"
-#include "Basis/Integrals/FunctionalOneElectronPrimitiveIntegralEngine.hpp"
-#include "Basis/Integrals/FunctionalTwoElectronPrimitiveIntegralEngine.hpp"
-#include "Basis/Integrals/HermiteCoulombIntegral.hpp"
 #include "Basis/Integrals/IntegralCalculator.hpp"
 #include "Basis/Integrals/IntegralEngine.hpp"
 #include "Basis/Integrals/Interfaces/LibcintInterfacer.hpp"
@@ -37,17 +34,20 @@
 #include "Basis/Integrals/Interfaces/LibintOneElectronIntegralEngine.hpp"
 #include "Basis/Integrals/Interfaces/LibintTwoElectronIntegralBuffer.hpp"
 #include "Basis/Integrals/Interfaces/LibintTwoElectronIntegralEngine.hpp"
-#include "Basis/Integrals/McMurchieDavidsonCoefficient.hpp"
 #include "Basis/Integrals/OneElectronIntegralBuffer.hpp"
 #include "Basis/Integrals/OneElectronIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveAngularMomentumIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveCanonicalKineticEnergyIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveCartesianOperatorIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveCoulombRepulsionIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveDipoleIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveLinearMomentumIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveNuclearAttractionIntegralEngine.hpp"
-#include "Basis/Integrals/PrimitiveOverlapIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/FunctionalOneElectronPrimitiveIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/FunctionalTwoElectronPrimitiveIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/HermiteCoulombIntegral.hpp"
+#include "Basis/Integrals/Primitive/McMurchieDavidsonCoefficient.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveAngularMomentumIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveCanonicalKineticEnergyIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveCartesianOperatorIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveCoulombRepulsionIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveDipoleIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveLinearMomentumIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveNuclearAttractionIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveOverlapIntegralEngine.hpp"
 #include "Basis/Integrals/TwoElectronIntegralBuffer.hpp"
 #include "Basis/Integrals/TwoElectronIntegralEngine.hpp"
 #include "Basis/MullikenPartitioning/GMullikenPartitioning.hpp"
