+++ gqcp/src/Basis/Integrals/CMakeLists.txt
@@ -1,13 +1,2 @@
-target_sources(gqcp
-    PRIVATE
-        BoysFunction.cpp
-        HermiteCoulombIntegral.cpp
-        IntegralEngine.cpp
-        McMurchieDavidsonCoefficient.cpp
-        PrimitiveAngularMomentumIntegralEngine.cpp
-        PrimitiveCartesianOperatorIntegralEngine.cpp
-        PrimitiveDipoleIntegralEngine.cpp
-        PrimitiveLinearMomentumIntegralEngine.cpp
-)
-
 add_subdirectory(Interfaces)
+add_subdirectory(Primitive)
