+++ gqcp/src/Basis/Integrals/Primitive/HermiteCoulombIntegral.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "Basis/Integrals/HermiteCoulombIntegral.hpp"
+#include "Basis/Integrals/Primitive/HermiteCoulombIntegral.hpp"
 
 #include "Mathematical/Functions/BoysFunction.hpp"
