+++ gqcp/src/Basis/Integrals/Primitive/PrimitiveCartesianOperatorIntegralEngine.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "Basis/Integrals/PrimitiveCartesianOperatorIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/PrimitiveCartesianOperatorIntegralEngine.hpp"
 
 #include <stdexcept>
