+++ gqcp/src/Mathematical/Functions/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(gqcp
     PRIVATE
+        BoysFunction.cpp
         CartesianExponents.cpp
         CartesianGTO.cpp
         LondonCartesianGTO.cpp
