+++ gqcpy/src/Basis/Integrals/HermiteCoulombIntegral_bindings.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "Basis/Integrals/HermiteCoulombIntegral.hpp"
+#include "Basis/Integrals/Primitive/HermiteCoulombIntegral.hpp"
 
 #include <pybind11/eigen.h>
 #include <pybind11/pybind11.h>
