+++ gqcpy/src/Basis/Integrals/IntegralCalculator_bindings.cpp
@@ -15,10 +15,10 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "Basis/Integrals/FunctionalOneElectronPrimitiveIntegralEngine.hpp"
-#include "Basis/Integrals/FunctionalTwoElectronPrimitiveIntegralEngine.hpp"
 #include "Basis/Integrals/IntegralCalculator.hpp"
 #include "Basis/Integrals/OneElectronIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/FunctionalOneElectronPrimitiveIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/FunctionalTwoElectronPrimitiveIntegralEngine.hpp"
 #include "Basis/Integrals/TwoElectronIntegralEngine.hpp"
 #include "Basis/ScalarBasis/GTOShell.hpp"
 #include "Basis/ScalarBasis/ShellSet.hpp"
