+++ gqcpy/src/Basis/Integrals/OneElectronIntegralEngine_bindings.cpp
@@ -15,8 +15,8 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "Basis/Integrals/FunctionalOneElectronPrimitiveIntegralEngine.hpp"
 #include "Basis/Integrals/OneElectronIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/FunctionalOneElectronPrimitiveIntegralEngine.hpp"
 #include "Utilities/aliases.hpp"
 
 #include <pybind11/pybind11.h>
