+++ gqcpy/src/Basis/Integrals/TwoElectronIntegralEngine_bindings.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "Basis/Integrals/FunctionalTwoElectronPrimitiveIntegralEngine.hpp"
+#include "Basis/Integrals/Primitive/FunctionalTwoElectronPrimitiveIntegralEngine.hpp"
 #include "Basis/Integrals/TwoElectronIntegralEngine.hpp"
 #include "Utilities/aliases.hpp"
